“修改混用有符号数和无符号数”

+++ frameworks/core/components_ng/pattern/waterflow/layout/top_down/water_flow_segmented_layout.cpp
@@ -157,7 +157,7 @@ int32_t WaterFlowSegmentedLayout::CheckDirtyItem() const
 {
     auto props = DynamicCast<WaterFlowLayoutProperty>(wrapper_->GetLayoutProperty());
     for (int32_t i = info_->startIndex_; i <= info_->endIndex_; ++i) {
-        if (info_->itemInfos_.size() <= i) {
+        if (static_cast<int32_t>(info_->itemInfos_.size()) <= i) {
             break;
         }
         float userDefHeight = WaterFlowLayoutUtils::GetUserDefHeight(sections_, info_->GetSegment(i), i);
